Feat: fixed PowerPlay 2.0 material name
Main change:
The names of the powerplay materials in the game have a code name that differs from what players see. Through the "Shiplocker.json" file I was able to trace almost all the in-game names of the various PP2.0 materials
- 3 materials still not updated:
- Power Mega Ship Data
- Energy Regulator
- Research Notes

Other changes
* Changed the comment with the name visible in the game
* Improved positioning and organization
* Moved some PP2.0 materials from the Data category to Item, previously set as such taking as source inara. However the Shiplocker.json file showed me their real category.

Why this change:
Without this correction they will not be identified by the program, this is because the code name and the one in the json do not match.